simplify module structure
